+++ tests/common/load_text_test.py
@@ -1,4 +1,4 @@
-from penelope.common.render_text import LemmaCorpusLoader, TaggedCorpusLoader, ZipLoader, ZippedTextCorpusLoader
+from penelope.corpus.render import LemmaCorpusLoader, TaggedCorpusLoader, ZipLoader, ZippedTextCorpusLoader
 from penelope.corpus.transforms import dehyphen
 
 # pylint: disable=protected-access
