+++ tests/notebook/topic_modelling/find_topic_documents_gui_test.py
@@ -1,7 +1,7 @@
 import pandas as pd
 
 from penelope import topic_modelling as tm
-from penelope.common import render_text as rt
+from penelope.corpus import render as rt
 from penelope.corpus import load_document_index
 from penelope.notebook import topic_modelling as ntm
